application: command-line plugin spec may be disabled, add SinglePluginApplication
In MainApplication the command-line plugin specification can be disabled
by the new 'disable_plugins_from_cmdline' option.

The feature is complete with the new SinglePluginApplication application
which has two options: the application name (program_name) and the plugin
name to be loaded - it's fallback_to_plugin_name in MainApplication.

The usage is much simpler and straightforward with the new class.