Application: add support of DEWI_DEBUG env var, if 1, same as '--debug'
If the SingleCommandApplication is in use, the app name + '_ARGS' is
the env var to pass parameters to the Application object,
but mostly it is complicated and only the debug is needed (for me),
thus the DEWI_DEBUG=1 is a straightforward solution, independent
from the usage of DEWI core's different *Application classes.

For the record the other solution is eg. for app name 'my-app-name':
export MY_APP_NAME_ARGS='-d'.