Update code based on pylint, codestyle except a refactor in appliction.py
Any global constant can be similar to class name ending with Plugin.
Any attribute can be in range ^[k-z]$ as it is realtively common.
For simplicity disable check of global statement.
The _r is a valid method name (for Plugin, but added to the .pylintrc).
One more refactor is needed in Application to reach the 10pt by pylint,
but not yet done.