Extract format/print of exception tracebacks, context/cause
The context: during handling the exception another exception is occurred
The cause: the exeption (a) is the direct cause of another (b),
 so in the code: raise b from a

All of these are printed from now on similar to the builtin
python codes.

As this may be useful anywhere else, moved into dewi_core.utils.exception.
By default sys.exc_info is used, but as a parameter any TracebackException
can be passed.