Add support of remoting: (de)serialize argparse.Namespace, run cmd
It's possible to run the same command on a remote host, eg.
by serializing argparse.Namespace to the stdin of an SSH command.
The serialization and the deserialization, and also the run of the command
requires knowledge of internal details which may be changed over time.
So the internals are hidden by the 'remoting' module and
the appplication.py's run_command_from_plugin_ns() method.