Add (prepare) support of XML-based man pages.
In the '00s the git used XML-based man pages, and I created a similar
functionality with the support of inclusion of common man page parts,
eg. the 'help', 'debug' and 'logging' options or other app-specific parts.

Based on that code I recreated the functionality for DEWI. While
the generator itself should work, the actual parameters and some other
parts of the process are uncertain yet, so it's not yet wired in.