command_registry got the functionality of registrar (register_class)
CommandRegistry is too complex, it's interface is almost unused in DEWI or
in my other projects, except the registrar's register_class, check if
a command is registered, and get the command class.

Because of this I added CommandRegistrar earlier, but it complicated the
code even further. To solve the issue I keep the Registrar as of now,
but it's wrapper code is moved into the CommandRegistry.

The CommandRegistry can be used everywhere without the CommandRegistrar.