The Application is simplified, plugin loading is external
The Application has two mode:
 * single command (literally, it's just a wrapper of a command)
 * multiple commands where command registry, listing commands, etc. are used

The plugin support is removed in that way, but still kept in another way:
Application.load_plugin() can load a plugin and register the commands into
the Application's command registry.

As a result there is only one Application, the SingleCommandApplication
and others are unneded and removed, including the 'miniapp' module.