Add Rails-like environment and config dir support, env var is DEWI_ENV
It can be useful to have separate environments for different use cases,
just like in Ruby on Rails. This is an optional feature, initially
it's unused.

Multiple config directories can be specified which are stored in the passed
Namespace object (config_directories_) as a list.

Each config dir can have an 'environment.py' file and a directory 'environments'
like Rails' config/environments/. The 'enviroments' can contain '<ENVNAME>.py'
and these are loaded in the following order:
- try to load from the config dirs in the order the config dirs added to the app
   -- in each config dir always load the environments.py
   -- load the .py file with the same name as the currently specified env
      in each config dir's environments/ dir (enviroments/<ENVNAME>.py)

- the env can be specified via the DEWI_ENV or the '-e' option.