Rewrite code using click
As a result the Command class is simpler, Subcommand is not needed.
Some parts are still in transient state, but the Application itself
works as before.

The Command.register_arguments() is either satic or class method,
it doesn't need the class, but it must be callable as
SomeCommand.register_arguments. It usies ClickHelper object to store
the options (see click.option, and so on; click_option_group for actual
params).

There is a badly implemented check of UTF-8 in click which is kicked out,
it prevented the test_application's tests to pass.

Added a few example commands, 'edit' is the same as the dewi_commands' edit,
and review, etc. to test multicommand setup including subcommands.

The Command.run() receives a ApplicationContext object instead of
argparse.Namespace, and ctx.current_args contain the current options/args,
but many other detail is also reachable without the magical 'name_'
form.

Remoting and README.md and some other parts are outdated.