Application: add optional description parameter
If the app is plugin-based, this is the only option to have a description,
or as click package uses: help.

If the app is in single command mode, the new description parameter
has priority over the command's description, so if the app desc parameter
is empty, the command's description is used, otherwise the app's.